♻️ Migrate the abstraction for the external service *Slack* into a single module.
